hack/ibm/requirements.txt
aiohttp==3.9.1
aiolimiter==1.1.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
httpx-sse==0.4.0
ibm-generative-ai==0.6.1
idna==3.6
multidict==6.0.4
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==1.26.18
yarl==1.9.4
